2006-04-05	Benedikt Meurer <benny@xfce.org>
	* configure.in.in: Do not depend on Xfce 4.2.2, as 4.2.1 will also do.
	* configure.in.in: Solaris has MD5 functions in libmd5.
	* exo/exo-toolbars-editor-dialog.c
	  (exo_toolbars_editor_dialog_add_toolbar): Fix format warning.
	* exo/exo-xsession-client.c(exo_xsession_client_set_group): Use
	  memcpy() instead of bcopy().
	* docs/reference/exo-open.xml: Fix typo.




(Old svn revision: 20759)

+++ exo/exo-toolbars-editor-dialog.c
@@ -153,7 +153,7 @@ exo_toolbars_editor_dialog_add_toolbar (ExoToolbarsEditorDialog *dialog)
   model = exo_toolbars_editor_get_model (EXO_TOOLBARS_EDITOR (dialog->priv->editor));
   if (G_LIKELY (model != NULL))
     {
-      name = g_strdup_printf ("exo-toolbar-%d-%d", getpid (), (gint) time (NULL));
+      name = g_strdup_printf ("exo-toolbar-%ld-%d", (glong) getpid (), (gint) time (NULL));
       exo_toolbars_model_add_toolbar (model, -1, name);
       g_free (name);
     }
